Moved utility classes for the SAML functional tests to a different package
